fix build errors

+++ src/mmchecking_list.cpp
@@ -735,7 +735,9 @@ void TransactionListCtrl::onMouseRightClick(wxMouseEvent& event)
     unsigned long col_nr = getColumnFromPosition(event.GetX());
     int flags;
     unsigned long row = HitTest(event.GetPosition(), flags);
-    if (row < m_trans.size() && (flags & wxLIST_HITTEST_ONITEM) && col_nr < getColNrSize()) {
+    if (row < m_trans.size() && (flags & wxLIST_HITTEST_ONITEM) &&
+        col_nr < static_cast<int>(getColNrSize())
+    ) {
         int col_id = getColId(col_nr);
         wxString menuItemText;
         wxString refType = Model_Attachment::REFTYPE_NAME_TRANSACTION;
